execution/example_script.py
#!/usr/bin/env python3
"""
Example Execution Script
Purpose: Template for creating deterministic execution scripts
Author: Auto-generated
"""

import argparse
import logging
from pathlib import Path
from dotenv import load_dotenv
import os

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)


def process_file(input_file: Path, output_file: Path) -> None:
    """
    Process the input file and generate output.
    
    This is a template function - replace with actual logic.
    
    Args:
        input_file: Path to input file
        output_file: Path to output file
        
    Raises:
        FileNotFoundError: If input file doesn't exist
        ValueError: If input file is invalid
    """
    if not input_file.exists():
        raise FileNotFoundError(f"Input file not found: {input_file}")
    
    logger.info(f"Processing file: {input_file}")
    
    try:
        # Your processing logic here
        with open(input_file, 'r') as f:
            data = f.read()
        
        # Process data...
        result = data  # Replace with actual processing
        
        # Write output
        output_file.parent.mkdir(parents=True, exist_ok=True)
        with open(output_file, 'w') as f:
            f.write(result)
        
        logger.info(f"Output written to: {output_file}")
        
    except Exception as e:
        logger.error(f"Error processing file: {e}")
        raise


def main() -> None:
    """Main entry point for the script."""
    parser = argparse.ArgumentParser(
        description="Example execution script template"
    )
    parser.add_argument(
        "--input",
        type=Path,
        required=True,
        help="Input file path"
    )
    parser.add_argument(
        "--output",
        type=Path,
        required=True,
        help="Output file path"
    )
    parser.add_argument(
        "--verbose",
        action="store_true",
        help="Enable verbose logging"
    )
    
    args = parser.parse_args()
    
    if args.verbose:
        logger.setLevel(logging.DEBUG)
    
    try:
        process_file(args.input, args.output)
        logger.info("Script completed successfully")
    except Exception as e:
        logger.error(f"Script failed: {e}")
        exit(1)


if __name__ == "__main__":
    main()
